pyproject.toml
[project]
name = "health-wellness-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "plotly>=5.18.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "reportlab>=4.0.7",
    "pydantic>=2.5.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "openai>=1.3.0"
]
